# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "初始化路由",
        "title": "API",
        "contact": {
            "name": "wty92911",
            "url": "https://github.com/wty92911",
            "email": "2066946886@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/family": {
            "get": {
                "description": "获取所有家庭的列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "获得所有家庭",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Family"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/family/create": {
            "post": {
                "description": "创建一个新的家庭",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "创建家庭",
                "parameters": [
                    {
                        "type": "string",
                        "description": "家庭名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Family"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/family/details": {
            "get": {
                "description": "获取指定家庭的详细信息，包括可选的预加载项",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "获得家庭详细情况",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "家庭ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "预加载项",
                        "name": "preloads",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Family"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/family/join": {
            "post": {
                "description": "加入一个现有的家庭",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "加入家庭",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "家庭ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Family"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/file/{url}": {
            "delete": {
                "description": "根据文件路径删除文件",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件路径",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/upload": {
            "post": {
                "description": "上传文件",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件路径",
                        "name": "path",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "github_com_wty92911_GoPigKit_internal_model.Category": {
            "type": "object",
            "required": [
                "family_id",
                "image_url",
                "mid_name",
                "name",
                "top_name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "family_id": {
                    "type": "integer"
                },
                "foods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Food"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "mid_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "top_name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_wty92911_GoPigKit_internal_model.Family": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Category"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "menu_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.MenuItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Order"
                    }
                },
                "owner": {
                    "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.User"
                },
                "owner_open_id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.User"
                    }
                }
            }
        },
        "github_com_wty92911_GoPigKit_internal_model.Food": {
            "type": "object",
            "required": [
                "category_id",
                "created_by",
                "image_urls",
                "price",
                "title"
            ],
            "properties": {
                "category_id": {
                    "description": "食品分类ID，外键",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdUser": {
                    "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.User"
                },
                "created_by": {
                    "description": "创建者",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "desc": {
                    "description": "食品描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_urls": {
                    "description": "食品图片",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "description": "食品价格",
                    "type": "integer"
                },
                "title": {
                    "description": "食品名称",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_wty92911_GoPigKit_internal_model.MenuItem": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "family_id": {
                    "description": "所属家庭ID",
                    "type": "integer"
                },
                "food_id": {
                    "description": "食品ID",
                    "type": "integer"
                },
                "quantity": {
                    "description": "数量",
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.User"
                }
            }
        },
        "github_com_wty92911_GoPigKit_internal_model.Order": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "family_id": {
                    "description": "家庭ID，外键",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "description": "关联的订单项，联级删除",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.OrderItem"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_wty92911_GoPigKit_internal_model.OrderItem": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "创建者，外键",
                    "type": "string"
                },
                "food_id": {
                    "description": "食品ID",
                    "type": "integer"
                },
                "order_id": {
                    "description": "订单ID，外键",
                    "type": "integer"
                },
                "quantity": {
                    "description": "数量",
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.User"
                }
            }
        },
        "github_com_wty92911_GoPigKit_internal_model.User": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "family_id": {
                    "description": "user创建时可以不指定family_id，等后面再加入一个家庭",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "open_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}